+++ data-filter/lib/filter/controllers/filter.controller.ts
@@ -63,7 +63,7 @@ export class FilterController<Data> {
 
     @Get("filter/config")
     public async getFilterConfig(@Req() req: any): Promise<FilterConfigurationModel[]> {
-        return await this.filterService.getConfig(await this.getUser(req) ?? {} as DataFilterUserModel);
+        return await this.filterService.getConfig(req, await this.getUser(req) ?? {} as DataFilterUserModel);
     }
 
     @Get("filter/config/id")
@@ -71,7 +71,7 @@ export class FilterController<Data> {
         @Query() search: FilterResourceValueModel,
         @Req() req: any
     ): Promise<SelectFilterValue> {
-        return await this.filterService.findResourceValueById(search, await this.getUser(req) ?? {} as DataFilterUserModel);
+        return await this.filterService.findResourceValueById(search, req, await this.getUser(req) ?? {} as DataFilterUserModel);
     }
 
     @Get("filter/config/value")
@@ -79,7 +79,7 @@ export class FilterController<Data> {
         @Query() search: FilterConfigurationSearchModel,
         @Req() req: any
     ): Promise<SelectFilterValue[]> {
-        return await this.filterService.searchConfigValues(search, await this.getUser(req) ?? {} as DataFilterUserModel);
+        return await this.filterService.searchConfigValues(search, req, await this.getUser(req) ?? {} as DataFilterUserModel);
     }
 
     protected async getUser(req: any): Promise<DataFilterUserModel> {