+++ data-filter/lib/filter/filters/select.filter.ts
@@ -9,17 +9,29 @@ export interface SelectFilterValue {
     name: string;
 }
 
+export interface SelectFilterGetValuesOptions<T = any, User = any, Request = any> {
+    value: T;
+    user?: User;
+    req?: Request
+}
+
+export interface SelectFilterGetResourceOptions<User = any, Request = any> {
+    id: number | string;
+    user?: User;
+    req?: Request
+}
+
 export interface SelectFilterDefinition<T> {
-    values: (value: unknown, user?: any) => Promise<SelectFilterValue[]>;
+    values: ({ value, user, req }: SelectFilterGetValuesOptions) => Promise<SelectFilterValue[]>;
     lazyLoading?: boolean;
-    getResourceById?: (id: number | string, user?: any) => Promise<SelectFilterValue>;
+    getResourceById?: ({ id, user, req }: SelectFilterGetResourceOptions) => Promise<SelectFilterValue>;
 }
 
 export class SelectFilter<T> extends Filter implements SelectFilterDefinition<T> {
     public type = FilterType.Select;
     public operators = [...SelectOperators];
-    public values: (value: unknown, user?: any) => Promise<SelectFilterValue[]>;
-    public getResourceById: (id: number | string, user?: any) => Promise<SelectFilterValue>;
+    public values: ({ value, user, req }: SelectFilterGetValuesOptions) => Promise<SelectFilterValue[]>;
+    public getResourceById: ({ id, user, req }: SelectFilterGetResourceOptions) => Promise<SelectFilterValue>;
     public lazyLoading;
 
     constructor(definition: BaseFilterDefinition & SelectFilterDefinition<T>) {
@@ -30,15 +42,15 @@ export class SelectFilter<T> extends Filter implements SelectFilterDefinition<T>
         }
     }
 
-    public async getConfig(key: string, user?: DataFilterUserModel): Promise<FilterBaseConfigurationModel> {
+    public async getConfig<Request>(key: string, req: Request, user?: DataFilterUserModel): Promise<FilterBaseConfigurationModel> {
         if (this.private) {
             return;
         }
 
-        const config = await super.getConfig(key, user);
+        const config = await super.getConfig(key, req, user);
         return {
             ...config,
-            values: this.lazyLoading ? [] : await this.values(null, user),
+            values: this.lazyLoading ? [] : await this.values({ value: null, user, req }),
             lazyLoading: this.lazyLoading
         };
     }