+++ data-filter/CHANGELOG.md
@@ -23,3 +23,38 @@ All notable changes to the @recursyve/nestjs-data-filter project will be documen
 ### Changed
  
 - Condition in filter configuration now accept a function as a value to allow the runtime rule value to used instead of juste a *hardcoded* value
+
+## [8.1.3] - 2022-04-11
+
+### Changed
+ 
+- Update typings for SelectFilter getResourceById
+
+## [8.2.0] - 2022-04-12
+
+### Changed
+
+- Filter config can now receive the Http Request in parameters
+
+### Breaking changes
+ 
+- Update SelectFilter `values` and `getResourceById` definition to accept request in parameters
+
+#### Before
+```typescript
+SelectFilter({
+    id: "test",
+    values: (value, user) => Promise.resolve([]),
+    getResourceById: (id, user) => Promise.resolve({ id: "test", name: "TEST!" })
+})
+```
+
+#### After
+
+```typescript
+SelectFilter({
+    id: "test",
+    values: ({ value, user, req }) => Promise.resolve([]),
+    getResourceById: ({ id, user, req }) => Promise.resolve({ id: "test", name: "TEST!" })
+})
+```